server/requirements.txt
Django==1.10.1
django-filter==0.14.0
django-suit==0.2.21
djangorestframework==3.4.6
psycopg2==2.6.2
requests==2.11.1
wheel==0.24.0
